+++ src/main/java/com/sparta/ordering/payment/dto/PaymentResponse.java
@@ -4,6 +4,7 @@ import com.sparta.ordering.payment.entity.Payment;
 import com.sparta.ordering.payment.entity.PaymentMethod;
 import com.sparta.ordering.payment.entity.PaymentStatus;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -11,7 +12,7 @@ public record PaymentResponse(
         UUID paymentId,
         UUID orderId,
         PaymentMethod paymentMethod,
-        Long amount,
+        BigDecimal amount,
         PaymentStatus status,
         String cancelReason,
         Instant canceledAt,
